Created a text file containing list of countries together with their capital city
Created a text file and populated an array with countries and capital cities delimited by a dash. There is a function that populates the array/list and returns said list. This array will be used in a function which creates the text file and copies each element in the list inside the text file created, hence the use of the for loop was mandatory. In the main function, we called the function which creates the text files.