Created a method that accesses and returns files to avoid repetition of code
Function accessFile takes two parameters, that is the file you wish to access and level of access, be it write, read or append. The method createFile has been renamed to populateTextFile and no longer accepts parameters as we already passed them in the function which accesses and returns the file.